Added image and header section
I also changed the navigation from services for future reference as I'm actually making the site and I was about to put my expertises together with the about me section. But for now I'll just make a seperate section for the expertise and add the services later on in the future.